requirements.txt
flask>=3.1.0
numpy>=2.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
supabase>=2.0.0
tensorflow>=2.16.0
gunicorn>=22.0.0
